refactor: extract heater mixin from ICModelController (step 4)
Move heater/setpoint setters and body_supports_cooling into src/pyintellicenter/_mixins/heater.py, composed into ICModelController. Completes the god-class decomposition: controller.py now holds connection/coalescing core, the cross-cutting entity aggregators, and backward-compat constant re-exports. Behavior-preserving: public API and __all__ unchanged, signatures identical; controller-module constant namespace preserved (guarded by test_controller_namespace_compat).
