+++ src/pyintellicenter/controller.py
@@ -23,6 +23,7 @@ from ._mixins import (
     _ChemistryMixin,
     _CircuitGroupMixin,
     _CoverMixin,
+    _HeaterMixin,
     _LightMixin,
     _PumpMixin,
     _ScheduleMixin,
@@ -92,22 +93,7 @@ from .attributes import (
     ASSIGN_ATTR as ASSIGN_ATTR,
 )
 from .attributes import (
-    BODY_ATTR,
-    HEATER_TYPE,
-    HITMP_ATTR,
-    LOTMP_ATTR,
-    MODE_ATTR,
-    OBJTYP_ATTR,
-    PARENT_ATTR,
-    PROPNAME_ATTR,
-    SNAME_ATTR,
-    STATUS_ATTR,
-    STATUS_OFF,
-    STATUS_ON,
-    SUBTYP_ATTR,
-    SYSTEM_TYPE,
-    VER_ATTR,
-    HeaterType,
+    BODY_ATTR as BODY_ATTR,
 )
 from .attributes import (
     BODY_TYPE as BODY_TYPE,
@@ -140,10 +126,19 @@ from .attributes import (
     HEATER_ATTR as HEATER_ATTR,
 )
 from .attributes import (
+    HEATER_TYPE as HEATER_TYPE,
+)
+from .attributes import (
+    HITMP_ATTR as HITMP_ATTR,
+)
+from .attributes import (
     HTMODE_ATTR as HTMODE_ATTR,
 )
 from .attributes import (
     LIGHT_EFFECTS as LIGHT_EFFECTS,
+)
+from .attributes import (
+    LOTMP_ATTR as LOTMP_ATTR,
 )
 from .attributes import (
     MAX_ATTR as MAX_ATTR,
@@ -156,6 +151,19 @@ from .attributes import (
 )
 from .attributes import (
     MINF_ATTR as MINF_ATTR,
+)
+from .attributes import (
+    MODE_ATTR,
+    OBJTYP_ATTR,
+    PARENT_ATTR,
+    PROPNAME_ATTR,
+    SNAME_ATTR,
+    STATUS_ATTR,
+    STATUS_OFF,
+    STATUS_ON,
+    SUBTYP_ATTR,
+    SYSTEM_TYPE,
+    VER_ATTR,
 )
 from .attributes import (
     NULL_OBJNAM as NULL_OBJNAM,
@@ -240,6 +248,9 @@ from .attributes import (
 )
 from .attributes import (
     VALVE_TYPE as VALVE_TYPE,
+)
+from .attributes import (
+    HeaterType as HeaterType,
 )
 from .connection import DEFAULT_TCP_PORT, DEFAULT_WEBSOCKET_PORT, ICConnection, TransportType
 from .exceptions import ICCommandError, ICConnectionError, ICResponseError, ICTimeoutError
@@ -625,6 +636,7 @@ class ICModelController(
     _PumpMixin,
     _BodyMixin,
     _SystemMixin,
+    _HeaterMixin,
     ICBaseController,
 ):
     """Controller that maintains a PoolModel of equipment state."""
@@ -886,73 +898,6 @@ class ICModelController(
         await self._flush_pending_changes()
         return await request.future
 
-    async def set_heat_mode(self, body_objnam: str, mode: HeaterType) -> dict[str, Any]:
-        """Set the heat mode for a body of water.
-
-        Args:
-            body_objnam: Object name of the body (pool or spa)
-            mode: HeaterType enum value
-
-        Returns:
-            Response dictionary
-
-        Example:
-            await controller.set_heat_mode("B1101", HeaterType.HEATER)
-        """
-        return await self._queue_property_change(body_objnam, {MODE_ATTR: str(mode.value)})
-
-    async def set_setpoint(self, body_objnam: str, temperature: int) -> dict[str, Any]:
-        """Set the heating setpoint for a body of water.
-
-        This is the temperature the system will heat UP to.
-        Alias for set_heating_setpoint().
-
-        Args:
-            body_objnam: Object name of the body (pool or spa)
-            temperature: Target heating temperature (units match system config)
-
-        Returns:
-            Response dictionary
-        """
-        return await self._queue_property_change(body_objnam, {LOTMP_ATTR: str(temperature)})
-
-    async def set_heating_setpoint(self, body_objnam: str, temperature: int) -> dict[str, Any]:
-        """Set the heating setpoint for a body of water.
-
-        This is the temperature the system will heat UP to (LOTMP attribute).
-        For the cooling setpoint, use set_cooling_setpoint().
-
-        Args:
-            body_objnam: Object name of the body (pool or spa)
-            temperature: Target heating temperature (units match system config)
-
-        Returns:
-            Response dictionary
-
-        Example:
-            await controller.set_heating_setpoint("B1101", 84)
-        """
-        return await self._queue_property_change(body_objnam, {LOTMP_ATTR: str(temperature)})
-
-    async def set_cooling_setpoint(self, body_objnam: str, temperature: int) -> dict[str, Any]:
-        """Set the cooling setpoint for a body of water.
-
-        This is the temperature the system will cool DOWN to (HITMP attribute).
-        Only relevant for systems with heat pumps or chillers that support cooling.
-        The cooling setpoint must be higher than the heat setpoint.
-
-        Args:
-            body_objnam: Object name of the body (pool or spa)
-            temperature: Target cooling temperature (units match system config)
-
-        Returns:
-            Response dictionary
-
-        Example:
-            await controller.set_cooling_setpoint("B1101", 86)
-        """
-        return await self._queue_property_change(body_objnam, {HITMP_ATTR: str(temperature)})
-
     def _get_attr_as_int(self, objnam: str, attr: str) -> int | None:
         """Get an attribute value as an integer, or None if unavailable."""
         obj = self._model[objnam]
@@ -972,53 +917,6 @@ class ICModelController(
             except (ValueError, TypeError):
                 return None
         return None
-
-    def body_supports_cooling(self, body_objnam: str) -> bool:
-        """Check if a body has a heater that supports cooling.
-
-        UltraTemp heat pumps (SUBTYP="ULTRA") support both heating and cooling.
-        Gas heaters (SUBTYP="HEATER"), solar heaters (SUBTYP="SOLAR"), and
-        generic heaters (SUBTYP="GENERIC") do not support cooling.
-
-        This checks ALL heaters that support this body, not just the currently
-        active one, so it returns True even if the system is currently off or
-        using a different heater.
-
-        Args:
-            body_objnam: Object name of the body (pool or spa)
-
-        Returns:
-            True if any available heater for this body supports cooling
-
-        Example:
-            if controller.body_supports_cooling("B1101"):
-                # Show both heating and cooling setpoints
-                heat_sp = controller.get_body_heating_setpoint("B1101")
-                cool_sp = controller.get_body_cooling_setpoint("B1101")
-        """
-        body = self._model[body_objnam]
-        if not body:
-            _LOGGER.warning("body_supports_cooling: body %s not found", body_objnam)
-            return False
-
-        # Check ALL heaters to see if any support this body AND can cool
-        all_heaters = list(self._model.get_by_type(HEATER_TYPE))
-
-        for heater in all_heaters:
-            # Check if this heater supports this body
-            supported_bodies = heater[BODY_ATTR]
-            if supported_bodies:
-                body_list = supported_bodies.split(" ")
-                if body_objnam in body_list:
-                    # Check if this heater supports cooling via either:
-                    # 1. Subtype being ULTRA (UltraTemp heat pump)
-                    # 2. Having a COOL attribute set to "ON"
-                    has_ultra = heater.subtype == "ULTRA"
-                    has_cool = heater["COOL"] == "ON"
-                    if has_ultra or has_cool:
-                        return True
-
-        return False
 
     # =========================================================================
     # Entity Discovery Helpers (for Home Assistant integration setup)
